commit_graph: introduce edge types
Summary:
In this stack we are converting the commit graph algorithms to be generic over three possible graph edge types: the regular commit graph using parents, a cut down commit graph using the first parent only, and an expanded commit graph including subtree sources.

This diff introduces the `EdgeType` trait, and three different implementations depending on which graph you are interested in.

Reviewed By: andreacampi

Differential Revision: D86551369

fbshipit-source-id: bf9a2a94e479bf7ad6dc13439c7f5cca9b84ad00