Scoped trait implementation for WarmBookmarksCache
Summary:
### Background
The Bookmarks cache can be configured with various sets of warmers, depending on the specific data requirements of each service.

Currently, different services use different warmer sets, but the Bookmarks service only exposes the Hg set. As a result, these services cannot stop using the Bookmarks cache directly.

With this stack, we are extending the Bookmarks cache to track three distinct pointers: Hg, Git, and all configured warmers. The goal is for the Bookmarks service to configure all warmers and expose all three pointers through its API, enabling different clients to access the cache directly according to their needs.

### This change
Now we have the basic infrastructure to track the bookmarks, so we can implement the new scoped trait. This is still not being used, and assumes that the CS pointers are actually tracked, which is not the case yet, that comes future diffs.

The old BookmarksTrait now will be implemented automatically if there is an implementation of the scoped version.

Reviewed By: gustavoavena

Differential Revision: D85769782

fbshipit-source-id: 46ed29168dc23257bdd8ba358d5829562205bf31