Update is_warm to track the scopes we support (hg / git)
Summary:
### Background
The Bookmarks cache can be configured with various sets of warmers, depending on the specific data requirements of each service.

Currently, different services use different warmer sets, but the Bookmarks service only exposes the Hg set. As a result, these services cannot stop using the Bookmarks cache directly.

With this stack, we are extending the Bookmarks cache to track three distinct pointers: Hg, Git, and all configured warmers. The goal is for the Bookmarks service to configure all warmers and expose all three pointers through its API, enabling different clients to access the cache directly according to their needs.

### This change
The is_warm function returned true when all the warmers were ready, now we want to track all our scopes. This diff still doesn't change behaviour, we keep all the checks so that they wait until both flags are true. This is the current behaviour, wait for all warmers to be ready.

Reviewed By: gustavoavena

Differential Revision: D85769786

fbshipit-source-id: 4ecf42e3392ef09254b84727c19c10469dc42dd7