Add LatestBookmarkEntries to track the latest derived and underived CS
Summary:
### Background
The Bookmarks cache can be configured with various sets of warmers, depending on the specific data requirements of each service.

Currently, different services use different warmer sets, but the Bookmarks service only exposes the Hg set. As a result, these services cannot stop using the Bookmarks cache directly.

With this stack, we are extending the Bookmarks cache to track three distinct pointers: Hg, Git, and all configured warmers. The goal is for the Bookmarks service to configure all warmers and expose all three pointers through its API, enabling different clients to access the cache directly according to their needs.

### This change
This diff adds a new struct so that we can keep track of the 3 bookmark entries, once the infrastructure is in place. For now we ignore the specific Hg / Git ones, and only track the one for all warmers, to match current behaviour.

Reviewed By: gustavoavena

Differential Revision: D85769785

fbshipit-source-id: 50f7c6f507a40d32bf7d881bc1c0a72fe92539a0