Add DerivedChangesetIds so that we can track the different changesets
Summary:
### Background
The Bookmarks cache can be configured with various sets of warmers, depending on the specific data requirements of each service.

Currently, different services use different warmer sets, but the Bookmarks service only exposes the Hg set. As a result, these services cannot stop using the Bookmarks cache directly.

With this stack, we are extending the Bookmarks cache to track three distinct pointers: Hg, Git, and all configured warmers. The goal is for the Bookmarks service to configure all warmers and expose all three pointers through its API, enabling different clients to access the cache directly according to their needs.

### This change
In the same way we did in the previous diff, we have a support struct to track the set of changesets we have per scope, but for now we only use the all tracking all warmers. As before, no behaviour is changing yet.

Reviewed By: gustavoavena

Differential Revision: D85769784

fbshipit-source-id: 0372f8359257a7b9b0ff775a50776df64fe49845