scmstore: get rid of some allocations of Vec<Key>
Summary:
There were a few spots creating a `Vec<Key>` of the pending keys unnecessarily. Change them to use `iter_pending` to avoid the intermediate allocation.

One spot lots the "total" value for its progress bar, but probably okay. We can pre-compute the count of pending items if we want it back.

Reviewed By: zzl0

Differential Revision: D87554309

fbshipit-source-id: 7d1d0a7df93c0e48d015f7041550e394ae83382b