prefetch: optimize memory usage of blob ids
Summary:
Add a new "stripObjectId" method that prefetch uses to save space on the saplng ObjectIds it accumulates. Basically, trim off the blob's path from the ObjectId leaving only the 20 byte node. The stripped ObjectId is short enough to fit in folly::fbstring's SSO range, which helps even more.

We don't need the path for anything when do a batch blob fetch. Walk detection uses the path in general, but ignores fetches from prefetching.

Note that I also tried to re-arrange prefetch so we kick off blob fetches as we go instead of accumulating one giant vector of ids. However this did not help because the tree traversal is so much faster than blob fetching that we end up holding all the ids in memory at the same time, anyway.

The new behavior is behind the experimental.prefetch-optimizations flag (which is enabled by default for tests).

Reviewed By: MichaelCuevas

Differential Revision: D87554303

fbshipit-source-id: 3880a467315208d947192ace8d978d2ae17ce441