scmstore: limit tree buffering
Summary: Limit the amount of trees and aux data we buffer before writing to the cache. It was previously unlimited because it is not normal to fetch lots of trees in a single query (because you need to traverse the trees to discover the ids to fetch). However, this assumption is making me nervous, so let's use some reasonable limits to ensure we don't use too much memory.

Reviewed By: zzl0

Differential Revision: D87554307

fbshipit-source-id: 85b006e691be8334d6ae1ad2402fc1fd769ed101