sql_ext: Fix broken tests
Summary:
As title.

One problem was a field missing in the parsing of the schematized logger.
Another one in the transaction test was that they started timing out getting a connection from the pool, so I increased the pool size from 1 to 2.

I also added identifiers to the assertions, so we can easily tell in failures if the issue is in the raw scuba logs or the schematized logger.

Reviewed By: chiara125

Differential Revision: D87640220

fbshipit-source-id: 2cba5c72628d66a9d42a97396012a81e4fc99d4a