add test that shows filter version mismatch issue
Summary:
# This diff

This demonstrates a bug in EdenMount::diff() where it compares RootIDs directly without considering that the FilterIds attached to each RootId have different versions. Even though the underlying commit hash is the same, changing the filter version creates a different FilterId, causing the parent commit mismatch error.

# Context

When FilteredFS is in use, each RootId comes with additional FilterId information "attached" to it. This extra information is used to consturct a FilteredRootId, which indicates the {RootId, Filter} pair that should be used to construct the root-tree of the repository. The FilterId indicates which identifier will be used to perform filtering on the Root object. This FilterId is then used to filter any subsequent trees that are fetched from the FilteredBackingStore.

The current code assumes that FilterIds are bijective, meaning we can easily confirm whether two FilteredRootIds refer to the same Root object by simply doing a bytewise comparison of the two FilteredRootIds. However, FilteredRootIds are not bijective in all cases. For example: two different RootIds can refer to the same HgSparseFilter Filter object. This is because HgSparseFilters include version information in them, and two FilterIds with different versions can refer to the same Filter.

Therefore, we need to do a more thorough check when comparing FilteredRootIds to determine if RootIds mismatch. We need to first check whether the underlying RootIds mismatch. If they don't, then we also need to confirm that the FilterIds don't mismatch (semantically, not bytewise). This stack implements the necessary logic to do semantic comparison of FilteredRootIds instead of bytewise comparison.

Differential Revision: D87029065

fbshipit-source-id: 54a65f67c6574351709f637069fe590b00e39260