EdenMount: fix RootId comparison
Summary:
# This diff

Uses the newly added compareRootsById to ensure that we can accurately compare two FilteredRootIds even if their FilterIDs differ bytewise, but are still semantically equal. Note: this does not fix the issue that would stem after this point, which is that all ObjectID comparisons would be super slow due to mismatching FilterIDs. I fix this in the following diffs in the stack.

# Context

See D87029065

Reviewed By: vilatto

Differential Revision: D87029074

fbshipit-source-id: ffe0781dab3aa13f789b50d91c534f090c8d605c