add filter equality method to Filter class
Summary:
# Context

Adds a new method to the base filter class that can be used to check if two filters are semantically the same. In most cases, this can simply be implemented as a bytewise comparison. However, some Filters may not be bijective, and therefore they may prefer to implement comparison differently.

I chose not to return ObjectComparison because I simply want to return whether the Filters are exactly the same or not. I don't care if they could potentially be the same/different.

# Context

See D87029065

Differential Revision: D87029075

fbshipit-source-id: b18f36a092ad54b94c3250499d3572c282acc608