sql_ext: Add more useful columns to the schematized logger
Summary:
## This stack
Making the last changes and improvements to the XDB telemetry schematized logger, to bring parity with the old raw scuba logging or add even more functionality.

## This diff
Log the mysql error number and error type. The error number will be used with a Scuba mapping column to get the Normalized error.

In the old scuba logs I derived that in Scuba, but it's better to have the columns in the schema, since we can use them in Hive too.

Differential Revision: D87647466

fbshipit-source-id: 88b10a312eb04f54b80eeb920441805b36f707cd