restricted_paths: fix tests that are flaky due to scuba logs
Summary:
The RestrictedPaths logging implementation constructs a scuba sample builder from new for each request, and uses an environment variable to pass the name of the file if we are diverting to a file in tests.  This is inherently racy, and causes problems in unit tests when they are run in parallel on multiple threads (as happens in OSS builds).

Switch instead to holding the scuba sample builder in the RestrictedPaths facet, and cloning from there each time we log.  This allows us to only read the env var once at start-up and also lets us override with different filenames in unit tests, which fixes the race.

Reviewed By: lmvasquezg

Differential Revision: D87739428

fbshipit-source-id: f7fe212884d3bee9cb1065e5dbc26647e5e899ca