Trigger cancellation on restart and shutdown
Summary:
In this diff, trigger cancellation on all active requests during shutdown, restart, and graceful restart.

Checkout is an example of an Eden request that if cancelled early, can leave the repo in a bad state. We want to give active requests the most time to complete. To accomplish this, the cancellation is triggered kCancellationOffsetBeforeShutdown (4) seconds before SIGKILL during receipt of a SIGTERM. As our confidence of cancellation improves, we can tighten this shorten this offset to be close to the SIGKILL time, to allow more time for requests to complete by themselves.

Reviewed By: MichaelCuevas, vilatto

Differential Revision: D87660568

fbshipit-source-id: 8e804012b91203a4e50dfcf097909008b5b91248