commit_rewriting: remove use of slog
Summary:
Remove all uses of slog in `commit_rewriting` modules.

These modules have utility functions to log to Scuba as well as `slog` or `tracing`.  This doesn't fit well with how scuba logging works, so we will likely change that in the future.  In the meantime, the `trace` variants never made it to scuba anyway as the log level would be at least `debug`, so switch those to regular `trace!` macro calls.

Reviewed By: lmvasquezg

Differential Revision: D78685434

fbshipit-source-id: 86dceb98ae46b1407613b2c25fd1df3ebabaca1d