Add multiple bookmark pointers
Summary:
### Background
The Bookmarks cache can be configured with various sets of warmers, depending on the specific data requirements of each service.

Currently, different services use different warmer sets, but the Bookmarks service only exposes the Hg set. As a result, these services cannot stop using the Bookmarks cache directly.

With this stack, we are extending the Bookmarks cache to track three distinct pointers: Hg, Git, and all configured warmers. The goal is for the Bookmarks service to configure all warmers and expose all three pointers through its API, enabling different clients to access the cache directly according to their needs.

### This change
This diff adds the tracking of the multiple bookmark pointers, using the infrastructure we added in the previous 7 diffs. It's adding the the tracking by using the previously ignored fields in the state, but it's not modifying the Bookmarks trait (and the new scoped trait is not in use yet), so nothing should change when this commit lands, users should see the same behaviour.

Now that the state is tracking the derived CS in the map, there is no need for the cs_id attribute anymore, so it's been removed to simplify the code.

Note that this shouldn't impact performance either, we are only looking at the configured warmers once, the same as before, so we are doing the same amount of work as before, but keeping track of 3 changesets instead of one.

Reviewed By: gustavoavena

Differential Revision: D85676372

fbshipit-source-id: a465cd2ca29ac128085407d1891ffb86886828d4