Consolidate BlobstorePutOps trait into Blobstore trait
Summary:
# This stack
I'm removing the `BlobstorePutOps` to simplify blobstore traits. The steps I'll follow are:
1. Remove BlobstorePutOps and make the methods part of Blobstore trait
2. Some code cleanup thanks to the reduction in code complexity
3. Implement the new methods where the implementation was missing (`put_with_status`  and `put_explicit`
4. Remove implementations of put, rely on default implementation always
5. Make `put_behaviour` arg to `put_explicit` optional
6. Make `put_with_status` have default implementation with all remaining logic for puts in put_explicit
7. Rename `put_Explicit` to put (potentially)

# This diff

This diff makes several changes to various files in the Mononoke codebase.

The primary change is the removal of the `BlobstorePutOps` trait from several functions, which is replaced by the `Blobstore` trait instead.

Additionally, new methods like `put_explicit` are added to `handle.rs` to support put operations with a specified behaviour. Similar `put_explicit` methods are also added to `mem_writes.rs`.

Overall, this diff refactors the blobstore functionality to remove the dependency on `BlobstorePutOps` and make it more flexible and generic by using the `Blobstore` trait instead.

Reviewed By: YousefSalama

Differential Revision: D87261238

fbshipit-source-id: b0544838a1c09db776b501e7f84452a43b6c3cf9