third-party/rust: point prettytable-rs at commit with unicode-width fix
Summary:
The unicode-width crate was pinned to a specific version in D64960800
to solve an issue with prettytable-rs. The prettytable-rs crate has not been
updated in a couple of years.

I published a branch on GitHub that contains prettytable-rs 0.10.0 + a fix for
the bug that lead to the pinning in
https://github.com/phsym/prettytable-rs/pull/170. Shipping this should enable
unpinning unicode-width. I am interested in this, because it's blocking an
import I am trying to do.

Note: it looks like projects in the Rust ecosystem are moving to the
comfy-table crate as a replacement for prettytable-rs. I might have a poke at
this too.

Reviewed By: dtolnay

Differential Revision: D88012122

fbshipit-source-id: 26d1caa67472c45a1480471b0745087301c58b1c