revise float32 constraint annotation
Summary: Revise the annotation `EnableUnsafeUnconstrainedFloat32` to generalize it for future customization of behavior around `inf` overflow, `NaN` acceptance, float32 truncation. Let users specify customizations. Currently, the behavior is `NONE`, but we expect to change the `truncation` and `inf_overflow` behavior to `MAP`. For now, we haven't implemented inf customization, so just use the `truncation` key to opt-out of the coming default change.

Reviewed By: aristidisp

Differential Revision: D87361193

fbshipit-source-id: f0406e462d165e55dac642a6098913ac5a470957