Fix for timing error in changesSince
Summary:
This diff fixes an issue where when calling getChangesSince, it could be possible to get a change from a sequence number 1 greater than the one returned as the latest sequence.

This is caused because at the top of the function, we record the latestJournalDelta, and later on call forEachDelta, which uses the current journalDelta value instead of the one recorded. If there has been a new entry in the journal since then, that value would be reported in forEachDelta.

This diff checks the sequenceID of the reported value, and then if it is greater than the current value, set the current value to that value.

Reviewed By: muirdm

Differential Revision: D87807713

fbshipit-source-id: 202316cf634edaad0e4d6c325afef186d6736c9a