fsck: migrate hg proxy hashes
Summary:
Make the fsck overlay checker detect hg proxy hashes as errors and repair them by rewriting the overlay entries with modern "embedded" object ids.

We attempt the migration if the clone is old (from 2022), we haven't done the migration before, and the fsck:migrate-proxy-hashes config is set (defaults true). If all these conditions are true, we "force" an fsck run on startup even if the overlay otherwise looks okay. This can be slow if you have an enourmous overlay, but it should only affect a small number of clones.

Recently I've been working on getting rid of the long deprecated hg proxy hashes. Hg proxy hashes refer to eden storing a mapping between its oid and hg's object hash and object path (in RocksDB). For various reasons this wasn't good, so we switched to embedding the hg info directly into eden's oid. However, since the hg proxy hashes can persist indefinitely in the overlay, we still need the code that supports reading them (and then resolving them via RocksDB). I'm removing the read-hg-proxy-hashes-code code, and I want to confidently not worry about the long tail of random old clones, so I'm going to forcibly migrate.

Reviewed By: genevievehelsel

Differential Revision: D87805800

fbshipit-source-id: 2f1b575782fb786df7abc2683d0e3141180b97e9