Add ability to ratelimit atlas
Summary:
In this diff I add the ability to ratelimit Atlas. For now we just check for `clientinfo_atlas` being set to True.
Later we will be able to include addtional metadata to check for in the struct, and we'll be able to extend the Thrift struct to match those

Note that for QPS ratelimiting we will be setting a single limit but it will apply to each main_client_id individually:

https://www.internalfb.com/code/fbsource/[7c2c14a89c569823b177ca7696b443340c7d022f]/fbcode/eden/mononoke/edenapi_service/src/middleware/rate_limiter.rs?lines=88

I've already set up atlas to get a different main_clientId in each environment: D87645412

Reviewed By: gustavoavena

Differential Revision: D87862487

fbshipit-source-id: c651f68acc0964d7c557e5ae055ab1640e50dd28