Remove recommended bookmarks GK
Summary:
The recommended bookmarks GK has been on 100% for good chunks of time (although disabled at the moment) and the feature is stable.

With the next diff, users who do not have a recommended bookmark will simply not see the recommended bookmarks section in the bookmarks manager. Essentially, having a recommended bookmark or not will be the new determiner if things are shown.

Removing the GK allows for lower latency on loading the commit tree on startup.

Reviewed By: evangrayk

Differential Revision: D86217927

fbshipit-source-id: 6d28b1d1b3d201753825e55b3cd37c5c1b3da659