renable github actions for linux (#1135)
Summary:
Pull Request resolved: https://github.com/facebook/sapling/pull/1135

Renable mononoke linux github actions so that folks externally can see that it current builds, and so they can see the breaks easily if it later doesn't build.

Added support for pinning rust version as during developement needed to pin to rust 1.90 as didn't yet compile with 1.91. In the meantime 1.91 support landed so pinned to 1.91.

Also saw that it was running out of disk space on the github actions runner so switched it to use MinSizeRel cmake build type and added a cargo mapping for it

Reviewed By: genevievehelsel

Differential Revision: D82818583

fbshipit-source-id: 6f0c54caa5b9593135a6977b1409f416ca622d81