Move eden/integration to Python 3.12
Summary:
With windows-curses now replaced with term-wiz, we should now be able to migrate eden/integration to 3.12 on Windows. Let's enable Python 3.12 for all platforms.

Moving to Python 3.12 will unblock the modern thrift-python migration.

There are a few compatibility warnings in the tests that I've hidden. Notably type warnings, and a deprecation warning on `forkpty`.

Reviewed By: vilatto

Differential Revision: D87949473

fbshipit-source-id: 975bd4ac6c12452fbf3776c5fda470b0b05d239c