change windowsRememberExecutableBit to true in EdenConfig
Summary:
I've opened this config to 100% of Windows users this week, and I am planning to remove this config from codebase next week. I want to change the default value of this config to true. Then we can confirm that with eden release next week everyone will have this config true regardless of what config manager is doing.
It helps me to confirm that removing this bool will not break anything

Reviewed By: muirdm

Differential Revision: D88406463

fbshipit-source-id: 227b3360e7ff0beda6eb4f0ae0f2f5c3297a5610