TreeInode: add "async" option when loading inodes
Summary:
Add an async param that causes the "start loading inode" logic to be guaranteed to run async (i.e. go through the executor rather than run sync). This includes both the ObjectStore::getTree call and the two loadOverlayDir calls in startLoadingInode.

We enable async=true for the `loadChildLocked()` codepath, which is what checkout/diff uses. The goal is to make sure the inode load is async for code that loads a bunch of inodes in a loop (i.e. we don't want to do heavy sync operations in a loop that kicks off 1000 inode loads). Other paths (i.e. from fs operations) use async=false since it isn't clear to me that forcing those "unary" fetches to be async is a good idea.

Note that currently execution is still not actually async because we still have `.via(QueuedImmediateExecutor)` which causes things to run inline in the current thread. My next diff will swap to a different executor here.

Like me, you may wonder why we need this `.via()`. The basic answer is that it is (currently) very important for the inode load to always complete. We don't want to return an ImmediateFuture or SemiFuture that the caller could end up dropping without scheduling. There are a couple options to handle this problem so we could switch to "normal" futures scheduled by caller (instead of a detached future), but they are more invasive/risky than this stack's approach.

The motivation for this change is to speed up "checkout" when the overlay (and potentially getTree) operations are slow/blocking. Previously we relied on the async point introduced by the local store for getTree calls (with store.async-rocksdb-local-store=true). With this diff the getTree and overload load calls are async on their own.

Note that the store.async-rocksdb-local-store=true flag only caused the getTree call to be async. The overlay loads were still not async with respect to the calling thread. So, this diff should give us even more asyncness than store.async-rocksdb-local-store=true.

Differential Revision: D88283123

fbshipit-source-id: 38171daf25323304f286d2874587cf1f5fc99f8e