checkout: pipe through checkout executor to TreeInode
Summary:
Pipe through the checkout executor to TreeInode::checkout (more specifically, the inode loading). If present, we use the executor instead of the QueuedImmediateExecutor when doing the tree fetching/overlay reading.

Effectively this activates the async points added in the previous diff. Take for example this "checkout" pseudo code:

```
futures = []
for child in tree:
  futures.push(child.startLoading())
collect(futures)
```

Previously this would often _serially_ fetch the child tree and/or load the child's overlay entry from disk. If your disk is slow, this makes checkout slow.

Now that we actually pass a thread pool through (and with the async points added by previous diff), the load operation is truly async with respect to the calling thread. This allows us to get a lot more parallelism during checkout.

Reviewed By: MichaelCuevas

Differential Revision: D88283121

fbshipit-source-id: 71e5123e1638a32faf77444f9906e410a446f66b