thrift: try to silence thread setpriority() warnings
Summary:
Switch from PRIORITY to SIMPLE thread manager type. I think/hope this will avoid thrift creating various priority thread pools. When it creates the "important" priority threads, it tries to de-nice them, but that fails (with an annoying warning in the logs) because we don't run as root or have the NICE capability.

The warning doesn't appear right at startup, so I'm not 100% sure this will fix it. It could be delayed action when the important threads are actually created/recycled, later.

Reviewed By: taidaii

Differential Revision: D88402146

fbshipit-source-id: 49e2462a5d377a7edeb471b97b06b286dc1a4beb