du: parallelize recursive size computation
Summary:
Slap some rayon in here to speed up the recursive size computation.

I looked at the jwalk library, but that only parallelizes the directory walking, not the file statting, so didn't go with it.

This speeds it up from 24s -> 4s on linux for 270GB directory.

On Windows I saw a similar speedup from 15s -> 3s (for a smaller directory).

Reviewed By: MichaelCuevas

Differential Revision: D88442163

fbshipit-source-id: 0df2ce493e637dc9df8dfc8b1c8e697cbe7b74d8