Add translation between derived data and bookmarks requirement
Summary:
The bookmarks config uses 'BookmarkCacheDerivedData' to represent the warmers that we need to apply to the bookmarks cache.

When using the remote version (or the scoped trait used by the bookmarks service), we have to specify the bookmarks requirement that we want to use. This defines how these enums need to be translated.

Note that No Derivation and Specific Types don't make sense when use in the remote cache. We will have checks to ensure that the are only used locally in an incoming diff.

Reviewed By: clara-9

Differential Revision: D87872349

fbshipit-source-id: 423aecadd44a22065bd5342729d39605702c7c95