separate platform specific invalidateChildrenNotMaterialized methods
Summary:
Create different `invalidateChildrenNotMaterialized` function for each platforms. It make it easier to read and understand what the code is doing, and we will need less number of `#ifdef`
As you can see in the code, the return parameters of the function is also different per platforms. Then it makes sense to have separate function per platform

Reviewed By: muirdm

Differential Revision: D86808667

fbshipit-source-id: 142c6f524deb1f2802f9881ac44e27ac6675c1fd