simpler log message
Summary: `numInvalidated` has different meaning on different platforms. It was used to show that GC is wroking. However, what is more important for GC is to show the number of inodes after GC is completed. In this stack I will remove this number later. However, for now we can simplify the log message and remove the `#ifdef` by just not printing the `numInvalidated`, and only print the number of inodes after GC to the log messages.

Reviewed By: vilatto

Differential Revision: D86786961

fbshipit-source-id: 1b8bfcc073a4d4e4ba5be0aa1525ecdb4c8bb7ca