extend GC integration tests to macOS platform
Summary:
We have integration tests on `invalidate_tets.py` for GC. The test was only availble on Windows because previously we only have GC on Windows.
Now that we have enabled GC on macOS, we should extend these integration tests to macOS platforms too.
The behaviour of GC on Windows and macOS is different.
On both platforms, both trees and files are invalidated, and deleted from memory. However, on macOS, only the invalidated trees are counted and reported by function, while invalidated files are not included in the count. These changes should reflect in the integration tests too.

In the next diff, I will remove some of these discrepency between two platforms and try to make the integration test same on both platforms

Differential Revision: D87012105

fbshipit-source-id: 44c4dc42e80f010eeeba2115fe4154f880f923df