Introduce fallback behavior for conf submit
Summary:
conf submit is not a perfect drop in for jf submit, which is much more flexible. Previously, we blocked all the places where conf submit was not compatible. This may lead to too much friction. Instead, we can go with a fallback approach. Use conf submit where applicable and fallback to jf submit otherwise so that it can be a transparent improvement for users. This will be a tradeoff in that we aren't blocking as much jf submits but in exchange we get less friction for end users.

This largely reverts D88221155 which disabled the different UI surfaces where submit would need to be blocked.

Reviewed By: evangrayk

Differential Revision: D88874940

fbshipit-source-id: 7fe3b942ed31bc3e2bec20df60e9960410696acb