commit_graph: don't filter common subtree source when writing prefetched commit graph
Summary:
The commit graph storage layer uses NULL for subtree source values to mean both "copy from the normal graph" and "has no value".  This is unfortunate in the case that the normal graph has a value but the subtree source graph should not, as it means we end up using the wrong value.  This stack addresses this by removing the fallback to the normal graph.

In this diff, we change it so that when writing out the prefetched commit graph, we don't filter out subtree source values that happen to correspond to the normal commit graph, as NULL is a valid value for the subtree source edges.

Reviewed By: chiara125

Differential Revision: D88444472

fbshipit-source-id: 6b761b2179b6974f5006f676d9120215e059e5be