commit_graph: don't fallback to parent edges when reading NULL subtree source edges
Summary:
The commit graph storage layer uses NULL for subtree source values to mean both "copy from the normal graph" and "has no value".  This is unfortunate in the case that the normal graph has a value but the subtree source graph should not, as it means we end up using the wrong value.  This stack addresses this by removing the fallback to the normal graph.

An earlier diff stopped writing NULL values to the database for subtree source edges that match parent edges.  Separately I have backfilled the DB for old NULLs and replaced them with the actual values. Now that we're not writing those NULL values, we should also stop falling back to parent edges when reading NULL values from the database. If a subtree source edge is NULL in the database, it should remain as None in the in-memory representation. This only applies to the edge fields (subtree_or_merge_ancestor, subtree_source_parent, subtree_source_skew_ancestor). The depth/gen fields still need their fallbacks for nowo since they can't be None and should never be NULL in the database.

Reviewed By: chiara125

Differential Revision: D88444469

fbshipit-source-id: 5c4de9df1784df1768a447b3d49243f621e6914e