commit_graph: apply subtree source fallback conditionally in preloaded storage
Summary:
The commit graph storage layer uses NULL for subtree source values to mean both "copy from the normal graph" and "has no value".  This is unfortunate in the case that the normal graph has a value but the subtree source graph should not, as it means we end up using the wrong value.  This stack addresses this by removing the fallback to the normal graph.

In this diff we update preloaded edges to conditionally apply subtree source fallback based on the scm/mononoke:commit_graph_disable_subtree_source_fallback JustKnob.

Reviewed By: chiara125

Differential Revision: D88444465

fbshipit-source-id: 2eb6cc22c3835a78d0d07e04f6f88b997a8d713d