HgProxyHash: drop support for reading from LocalStore
Summary:
We stopped writing the legacy proxy hash mappings to the LocalStore a long time ago. Let's clean up the read path. I'm working towards removing the LocalStore, and this is one of the remaining users.

Basically what I've done in this commit is move the embedded ObjectId parsing into HgProxyHash's constructor since it is now the only option (i.e. we no longer support the ObjectId containing a "proxy" hash that must be resolved using a LocalStore lookup).

Reviewed By: kavehahmadi60, MichaelCuevas

Differential Revision: D86211819

fbshipit-source-id: d20786f4fe960d1fc3dc7f81af8c823c2b4093f6