SaplingBackingStore: kill SlOid::getbatch
Summary:
`getBatch()` always copies the input ObjectIds into a vector of SlOids. Instead we can just create a vector of SlOidViews which avoids copying the underlying oid data for each oid.

The getBatch made more sense originally when it was resolving hg proxy hashes in the LocalStore. More recently it still did signifcant work because it did a per-id allocation (and there can be a lot of ids). But now we just have O(1) allocation of the SlOidView vector, so it doesn't seem necessary to have a separate getBatch or a separate async point.

Note that the "not optimized" code path in `SaplingBackingStore::prefetchBlobs()` still copies each SlOidView as it is passed to `enqueueBlob()` (but this is no different than previous behavior).

Reviewed By: vilatto

Differential Revision: D86263727

fbshipit-source-id: fc2198ed5cd102c36f1bdec55c465a0ffcef2052