SaplingObjectId: clean up proxy hash remnants
Summary:
Rework SaplingObjectId (previously HgProxyHash) to now be a simple container for the ObjectId data. Constructing a SaplingObjectId from ObjectId simply copies the data (and validates). Constructing a SaplingObjectId from sapling hash and path is stored directly in the ObjectId format.

Other random changes I'm making at the same time because I was in too deep:
- Rename `SaplingObjectId::revHash()` to `SaplingObjectId::node()`, and change return  value to `Hash20&` instead of `folly::ByteRange`. Although `Hash20` "limits" us to 20 byte hashes, that is what they are in practice, and `Hash20` is more useful since it formats automatically as a hex string, etc.

I think we already avoid some copies with this commit's simplification, but I'm going to follow up with a SaplingObjectIdView type to avoid copying `ObjectId&` for sync codepaths.

Reviewed By: vilatto, MichaelCuevas

Differential Revision: D86211811

fbshipit-source-id: 2850483f6434acc46b5b19762990850e79988484