mononoke_api: add support for additional changes when folding commits
Summary:
People currently use repo_create_commit to simulate "amend" and "squash" source control operations, but they are missing some useful features like mutation records.  This stack implements support for creating amended commits that are made up of one or more commits folded together with optional additional changes.

In this diff we add support for applying additional changes on top of the folded stack.  Note that we must implement the same checks as for creating a new commit to ensure that the additional changes result in a valid bonsai changeset.

Reviewed By: MichaelCuevas

Differential Revision: D89008408

fbshipit-source-id: adabe28658b404316145c9b602bec2aa659a2812