mononoke_api: added checks to prevent public commit folding
Summary:
People currently use repo_create_commit to simulate "amend" and "squash" source control operations, but they are missing some useful features like mutation records.  This stack implements support for creating amended commits that are made up of one or more commits folded together with optional additional changes.

In this diff we check that the commits being folded or amended are not public, as public commits should not be mutated.

Reviewed By: YousefSalama

Differential Revision: D87776662

fbshipit-source-id: 9e6ac7a640c94ede85da157c254fc416e8bb533f