reorg: Migrate scs to servers/scs/ (attempt #2)
Summary:
2nd attempt at D88852630.

Unlike the other binaries, SCS has a lot of reverse dependencies across the company's entire codebase, so the diff was quickly reverted and I didn't have time to fix forward.

This diff applies the same changes as D88852630 **but it does NOT move the thrift directory**. This stores the interface, not the implementation, so it should be fine to leave it there.

Reviewed By: markbt

Differential Revision: D88862754

fbshipit-source-id: da1a961d7ec3c9336c268ecce240accdb4aa4ea6