clearing FS ref count instead of decreasing during NFS GC
Summary:
On NFS, the FS refcount is a flag that could be 0 or 1. Because we don't have any FORGET on NFS to decrease FS refcount.

During NFS GC we decrease the FS ref count by 1 to delete the inode. However, it is cleaner to set `clearRefCount` and just make it zero instead of decreasing FS refcount.

This is same as what we did in the past for `InodeMap::forgetStaleInodes()` which is running for NFS and make the FS refcount zero to delete stale inodes during checkout.
see here:
https://www.internalfb.com/code/fbsource/[2cf99c7bf20c]/fbcode/eden/fs/inodes/InodeMap.cpp?lines=810-815

Reviewed By: genevievehelsel

Differential Revision: D88894840

fbshipit-source-id: 92844b00c774938407e2914116278c62f753d31e