cext/evalframe: split sapling_cext_evalframe_resolve_frame into multiple functions
Summary:
In the future, I'd like to be able to extract the minimal amount of "Python
stack" info (ex. the address to the code object, and the line number) first (in
a signal handler), then resolve its function and file names later (in a
separate thread that does not block main thread execution).

Split `sapling_cext_evalframe_resolve_frame` to do it.

Reviewed By: zzl0

Differential Revision: D88664593

fbshipit-source-id: 2cd50a340b1659a8a0b8d57248fb6fd758d8b333