Use subagents for large commits
Summary:
We are running into prompt size limitations when trying to split super large diffs.

Using subagents to process chunks of the commit data.

Also ask the agent to just provide ranges of line numbers instead of every single line number because it was also running into size limitations for the json. We can expand the line ranges after when actually applying the split in ISL

Differential Revision: D89201344

fbshipit-source-id: cf81d4db97d989708fcb1962d502e001a8271788