repos_manager: improve startup time with increased parallelism
Summary:
When we start up mononoke, we load all the repos in a single tokio task with `buffered(30)`.  This makes it really slow for a few reasons:

* Repos with lots of bookmarks stall the pipeline (the next 29 repos get loaded, but then we wait until that large one is complete before continuing).
* Running in a single task makes us vulnerable to the subexecutor problem where a slow repo may cause another repo to not get polled, causing timeouts.
* Running in a single task means we're only using a single CPU.

Fix all of these by moving to `buffer_unordered` (the order doesn't matter anyway), and spawning each repo load onto its own tokio task.  We'll increase
parallelism to 50 as well, while we are at it.

To make progress easier to understand, we also log how far we are through initialization when each repo completes.

Reviewed By: lmvasquezg, mzr

Differential Revision: D89373726

fbshipit-source-id: 20a9e790c9e2201bfa41bef6382c03fe47296503