Dedupe streaming states client and states client
Summary:
Context:

We split off edenfs-asserted-states-client‎ from edenfs-asserted-states due to needing a library to import without having a dependency on the rest of eden.

This diff:
Removes the duplicated code between the two. Leave the endpoints inside edenfs-asserted-states for now since I'm not sure which clients are currently using them.

Reviewed By: vilatto

Differential Revision: D89318974

fbshipit-source-id: 6d8e972f37fbe276d9b064ddd60db5237851198c