internalconfig: add helper for determining repo owner
Summary: In a future diff, we want to determine the unixname of the user that owns a specific path. We can easily do this on Linux/macOS, but it's not trivial to do on Windows. For now, we'll implement a way to find the owner of a file/dir on unix environments, and we'll no-op on Windows.

Reviewed By: muirdm

Differential Revision: D89401721

fbshipit-source-id: 82036c8fccc2b9c058dd42ea46224d87f0cfafeb