repos_manager: reduce concurrency from 50 to 30
Summary:
D89373726 changed to buffer_unordered and increased concurrency to 50.
We believe this could be causing the CPU spikes that are causing the health check failures in bookmark service: https://fburl.com/services/tgexjjbs

FYI I'll sl push this to create a hotfix release in conveyor now, bc bookmark service release is 14 days old and code freeze is starting soon.

jagg it might be worth exposing this in a JK so we can change it more quickly if needed.

Differential Revision: D89469378

fbshipit-source-id: ab3652ae09c2f5570a24b93bb2b8625b4cb057a9