Add cancellation token and throwIfCanceled to fetchContext
Summary:
To support EdenFS Thrift request cancellation on existing Futures based code, we need to manually plumb cancellation tokens through the call stack. An efficient way to do this is to add cancellation tokens to `ObjectFetchContext`. We can then check to return early based on this cancellation token with `throwIfCanceled`.

At V1 of this implementation, I had thought that `ThriftFetchContext` would be the only child class of `ObjectFetchContext` that would need to support cancellation. However, I have found that we want to support cancellation within the `CheckoutAction` which uses the `StatsFetchContext` -- thus I've move the cancellation support to the parent `ObjectFetchContext`

Current usage of `StatsFetchContext` makes a new context instead of using an existing `ObjectFetchContext`, thus we make changes to `CheckoutContext` and `DiffContext` to support plumbing the cancellation token.

Reviewed By: genevievehelsel, vilatto

Differential Revision: D88424773

fbshipit-source-id: 756a6d3f0ca3dc1101365c545d8a016e266c52c6